@mattersoflight 2023-02-28
@talonchandler 

>@Soorya19Pradeep asked about the consistency of this convention with the existing waveorder implementation.

Could you clarify which convention we are discussing here:  convention for what is called right vs left circularly polarized light, or convention for axes of orientation, or may be both?