@Soorya19Pradeep 2023-02-28
> @talonchandler
> 
> > @Soorya19Pradeep asked about the consistency of this convention with the existing waveorder implementation.
> 
> Could you clarify which convention we are discussing here: convention for what is called right vs left circularly polarized light, or convention for axes of orientation, or may be both?

@mattersoflight, I am trying to understand how to read the orientation measurement of cell membrane, if it makes physical sense. The value of orientation changes with new implementation and further background correction, so I was curious. 
This is from orientation information from a cell from earlier version with measured background correction, colorbar range of values (-0.87,+0.87)
<img width="444" alt="image" src="https://user-images.githubusercontent.com/101817974/221914487-29cf0229-ff9b-46a7-b26a-92bfb40dcfc7.png">
This is from the new version with just measured background correction, range (+0.87,+2.5). I realized the information here is same, just inverted and offset by 90 degrees.
<img width="444" alt="image" src="https://user-images.githubusercontent.com/101817974/221913601-0e56f4ac-e94c-40c6-a7ac-52fcffaa02d8.png">
After the extra background correction the range changes and more information is visible, range (0,+3.14).
<img width="444" alt="image" src="https://user-images.githubusercontent.com/101817974/221913869-0201bc9a-b4e4-46a5-8202-666240768d92.png">