@talonchandler 2023-02-28
Thanks for the comparison @Soorya19Pradeep. Very helpful. 

> I realized the information here is same, just inverted and offset by 90 degrees.

These are the types of knobs that we might provide in the "decoupling" approach: one checkbox/function for "invert orientation" and one for "rotate by 90 degrees".