@talonchandler 2023-03-07
I've just completed the renaming/refactoring. @mattersoflight this is ready for your re-review. 

The latest API (at this level) looks like:
```
# Calculate I2S
I2S = stokes.I2S_matrix(swing, scheme="5-State")

# Calculate Stokes vectors
S_sm = stokes.mmul(I2S, czyx_data)  # Apply I2S to sample data
S_bg_meas = stokes.mmul(I2S, cyx_bg_data)  # Apply I2S to measured background

# Calculate background correction matrix from S_bg
M_bg_inv = stokes.mueller_from_stokes(*S_bg_meas)

# Apply measured background correction to sample data
bg_corr_S_sm = stokes.mmul(M_bg_inv, S_sm)

# Recover ADR parameters
ret, ori, tra, dop = stokes.estimate_ADR_from_stokes(*bg_corr_S_sm)
```

I've also spent some time characterizing the old vs. new algorithms.

**Soorya's cells - retardance on y axis - measured bkg correction only**
![Screenshot 2023-03-06 at 5 53 03 PM](https://user-images.githubusercontent.com/9554101/223301263-a580ca6c-e835-4aaf-85cc-e4712134d70f.png)
At most 2% difference in retardance.

**Kazansky target - retardance on y axis - measured bkg correction only**
![Screenshot 2023-03-06 at 5 31 36 PM](https://user-images.githubusercontent.com/9554101/223301450-7e6e5638-03f3-4560-9273-76eeb8a59790.png)
At most 1% difference in retardance.

**Kazansky target - orientation on y axis - measured bkg correction only**
![Screenshot 2023-03-06 at 5 35 18 PM](https://user-images.githubusercontent.com/9554101/223301559-46b1ebab-9350-4c7e-90a1-3a1ed560b6ed.png)
At most 2% difference in non-background regions when the different orientation convention is accounted for. This main difference here is from a difference in orientation conventions which we'll be handling with two user-facing switches as discussed above. 

**Timing**
Current performance bottleneck is the pre-calculation of `mueller_from_stokes` from the background stokes vectors, which can be further optimized (I expect factors of 2-4x). For now, here are a couple benchmarks: 

1 x 2048 x 2048: 
old algorithm: 1.0 s 
new algorithm: 15.4 s 

8 x 2048 x 2048: 
old algorithm: 17.8 s 
new algorithm: 19.6 s 

**Example comparison script (generates Kaz target comparison above)**

<details>
  <summary>Full example script (click to expand):</summary>

```
import numpy as np
from waveorder import stokes
from recOrder.io.utils import load_bg
from recOrder.io.metadata_reader import MetadataReader
from recOrder.compute.reconstructions import (
    initialize_reconstructor,
    reconstruct_qlipp_stokes,
    reconstruct_qlipp_birefringence,
)
from iohub.reader import imread
from iohub.ngff import open_ome_zarr
import napari

# Set paths
base_path = "/hpc/projects/compmicro/rawdata/hummingbird/Talon/2023_02_08_kaz/"
data_subpath = "kaz-raw_recOrderPluginSnap_0/kaz-raw_RawPolDataSnap.zarr"
bg_subpath = "BG"
cal_subpath = "calibration_metadata.txt"

# Read data
reader = imread(base_path + data_subpath)
T, C, Z, Y, X = reader.shape
czyx_data = reader.get_array(position=0)[0, ...]

# Read background data
cyx_bg_data = load_bg(base_path + bg_subpath, height=Y, width=X)

# Read calibration metadata
md = MetadataReader(base_path + cal_subpath)

def new_bg_correction(czyx_data, cyx_bg_data, swing, scheme):
    # Calculate I2S
    I2S = stokes.I2S_matrix(md.Swing, scheme=md.get_calibration_scheme())

    # Calculate Stokes vectors
    S_sm = stokes.mmul(I2S, czyx_data)  # Apply I2S to sample data
    S_bg_meas = stokes.mmul(I2S, cyx_bg_data)  # Apply I2S to measured background

    # Calculate background correction matrix from S_bg
    M_bg_inv = stokes.mueller_from_stokes(*S_bg_meas)

    # Apply measured background correction to sample data
    bg_corr_S_sm = stokes.mmul(M_bg_inv, S_sm)

    # Recover ADR parameters
    ret, ori, tra, dop = stokes.estimate_ADR_from_stokes(*bg_corr_S_sm)

    ret = ret / (2 * np.pi) * 532

    return ret, ori, tra, dop

def old_bg_correction(czyx_data, cyx_bg_data, swing, scheme):
    reconstructor_args = {
        "image_dim": (Y, X),
        "n_slices": 1,  # number of slices in z-stack
        "wavelength_nm": 532,
        "swing": swing,
        "calibration_scheme": scheme,  # "4-State" or "5-State"
        "bg_correction": "global",
    }
    reconstructor = initialize_reconstructor(
        pipeline="birefringence", **reconstructor_args
    )
    # Reconstruct background Stokes
    bg_stokes = reconstruct_qlipp_stokes(cyx_bg_data, reconstructor)

    # Reconstruct data Stokes w/ background correction
    stokes = reconstruct_qlipp_stokes(czyx_data, reconstructor, bg_stokes)

    birefringence = reconstruct_qlipp_birefringence(stokes, reconstructor)
    birefringence[0] = (
        birefringence[0] / (2 * np.pi) * reconstructor_args["wavelength_nm"]
    )
    return birefringence

oldADR = old_bg_correction(czyx_data, cyx_bg_data, md.Swing, md.Calibration_scheme)
newADR = new_bg_correction(czyx_data, cyx_bg_data, md.Swing, md.Calibration_scheme)


v = napari.Viewer()
v.add_image(oldADR[..., 890:1220, 790:1370], name="old")
v.add_image(np.stack(newADR)[..., 890:1220, 790:1370], "new")
import pdb; pdb.set_trace()
```
</details>